refactor: split remaining monitor runtime helpers
