+++ extensions/telegram/src/monitor.test.ts
@@ -1,6 +1,7 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { monitorTelegramProvider } from "./monitor.js";
 import { tagTelegramNetworkError } from "./network-errors.js";
+
+type MonitorTelegramOpts = import("./monitor.js").MonitorTelegramOpts;
 
 type MockCtx = {
   message: {
@@ -136,6 +137,7 @@ function mockRunOnceAndAbort(abort: AbortController) {
 }
 
 async function expectOffsetConfirmationSkipped(offset: number | null) {
+  const { monitorTelegramProvider } = await import("./monitor.js");
   readTelegramUpdateOffsetSpy.mockResolvedValueOnce(offset);
   const abort = new AbortController();
   api.getUpdates.mockReset();
@@ -149,6 +151,7 @@ async function expectOffsetConfirmationSkipped(offset: number | null) {
 }
 
 async function runMonitorAndCaptureStartupOrder(params?: { persistedOffset?: number | null }) {
+  const { monitorTelegramProvider } = await import("./monitor.js");
   if (params && "persistedOffset" in params) {
     readTelegramUpdateOffsetSpy.mockResolvedValueOnce(params.persistedOffset ?? null);
   }
@@ -203,9 +206,8 @@ function expectRecoverableRetryState(expectedRunCalls: number) {
   expect(runSpy).toHaveBeenCalledTimes(expectedRunCalls);
 }
 
-async function monitorWithAutoAbort(
-  opts: Omit<Parameters<typeof monitorTelegramProvider>[0], "abortSignal"> = {},
-) {
+async function monitorWithAutoAbort(opts: Omit<MonitorTelegramOpts, "abortSignal"> = {}) {
+  const { monitorTelegramProvider } = await import("./monitor.js");
   const abort = new AbortController();
   mockRunOnceAndAbort(abort);
   await monitorTelegramProvider({
@@ -215,8 +217,8 @@ async function monitorWithAutoAbort(
   });
 }
 
-vi.mock("../../../src/config/config.js", async (importOriginal) => {
-  const actual = await importOriginal<typeof import("../../../src/config/config.js")>();
+vi.mock("openclaw/plugin-sdk/config-runtime", async (importOriginal) => {
+  const actual = await importOriginal<typeof import("openclaw/plugin-sdk/config-runtime")>();
   return {
     ...actual,
     loadConfig,
@@ -260,14 +262,22 @@ vi.mock("@grammyjs/runner", () => ({
   run: runSpy,
 }));
 
-vi.mock("../../../src/infra/backoff.js", () => ({
-  computeBackoff,
-  sleepWithAbort,
-}));
+vi.mock("openclaw/plugin-sdk/infra-runtime", async (importOriginal) => {
+  const actual = await importOriginal<typeof import("openclaw/plugin-sdk/infra-runtime")>();
+  return {
+    ...actual,
+    computeBackoff,
+    sleepWithAbort,
+  };
+});
 
-vi.mock("../../../src/infra/unhandled-rejections.js", () => ({
-  registerUnhandledRejectionHandler: registerUnhandledRejectionHandlerMock,
-}));
+vi.mock("openclaw/plugin-sdk/runtime-env", async (importOriginal) => {
+  const actual = await importOriginal<typeof import("openclaw/plugin-sdk/runtime-env")>();
+  return {
+    ...actual,
+    registerUnhandledRejectionHandler: registerUnhandledRejectionHandlerMock,
+  };
+});
 
 vi.mock("./webhook.js", () => ({
   startTelegramWebhook: startTelegramWebhookSpy,
@@ -282,6 +292,16 @@ vi.mock("./update-offset-store.js", () => ({
   writeTelegramUpdateOffset: vi.fn(async () => undefined),
 }));
 
+vi.mock("openclaw/plugin-sdk/reply-runtime", async (importOriginal) => {
+  const actual = await importOriginal<typeof import("openclaw/plugin-sdk/reply-runtime")>();
+  return {
+    ...actual,
+    getReplyFromConfig: async (ctx: { Body?: string }) => ({
+      text: `echo:${ctx.Body}`,
+    }),
+  };
+});
+
 vi.mock("../../../src/auto-reply/reply.js", () => ({
   getReplyFromConfig: async (ctx: { Body?: string }) => ({
     text: `echo:${ctx.Body}`,
@@ -292,6 +312,7 @@ describe("monitorTelegramProvider (grammY)", () => {
   let consoleErrorSpy: { mockRestore: () => void } | undefined;
 
   beforeEach(() => {
+    vi.resetModules();
     loadConfig.mockReturnValue({
       agents: { defaults: { maxConcurrent: 2 } },
       channels: { telegram: {} },
@@ -387,6 +408,7 @@ describe("monitorTelegramProvider (grammY)", () => {
   });
 
   it("retries on recoverable undici fetch errors", async () => {
+    const { monitorTelegramProvider } = await import("./monitor.js");
     const abort = new AbortController();
     const networkError = makeRecoverableFetchError();
     runSpy
@@ -410,6 +432,7 @@ describe("monitorTelegramProvider (grammY)", () => {
   });
 
   it("retries recoverable deleteWebhook failures before polling", async () => {
+    const { monitorTelegramProvider } = await import("./monitor.js");
     const abort = new AbortController();
     const cleanupError = makeRecoverableFetchError();
     api.deleteWebhook.mockReset();
@@ -423,6 +446,7 @@ describe("monitorTelegramProvider (grammY)", () => {
   });
 
   it("retries setup-time recoverable errors before starting polling", async () => {
+    const { monitorTelegramProvider } = await import("./monitor.js");
     const abort = new AbortController();
     const setupError = makeRecoverableFetchError();
     createTelegramBotErrors.push(setupError);
@@ -436,6 +460,7 @@ describe("monitorTelegramProvider (grammY)", () => {
   });
 
   it("awaits runner.stop before retrying after recoverable polling error", async () => {
+    const { monitorTelegramProvider } = await import("./monitor.js");
     const abort = new AbortController();
     const recoverableError = makeRecoverableFetchError();
     let firstStopped = false;
@@ -463,6 +488,7 @@ describe("monitorTelegramProvider (grammY)", () => {
   });
 
   it("stops bot instance when polling cycle exits", async () => {
+    const { monitorTelegramProvider } = await import("./monitor.js");
     const abort = new AbortController();
     mockRunOnceAndAbort(abort);
 
@@ -473,6 +499,7 @@ describe("monitorTelegramProvider (grammY)", () => {
   });
 
   it("clears bounded cleanup timers after a clean stop", async () => {
+    const { monitorTelegramProvider } = await import("./monitor.js");
     vi.useFakeTimers();
     try {
       const abort = new AbortController();
@@ -487,6 +514,7 @@ describe("monitorTelegramProvider (grammY)", () => {
   });
 
   it("surfaces non-recoverable errors", async () => {
+    const { monitorTelegramProvider } = await import("./monitor.js");
     runSpy.mockImplementationOnce(() =>
       makeRunnerStub({
         task: () => Promise.reject(new Error("bad token")),
@@ -497,6 +525,7 @@ describe("monitorTelegramProvider (grammY)", () => {
   });
 
   it("force-restarts polling when unhandled network rejection stalls runner", async () => {
+    const { monitorTelegramProvider } = await import("./monitor.js");
     const abort = new AbortController();
     const { stop } = mockRunOnceWithStalledPollingRunner();
     mockRunOnceAndAbort(abort);
@@ -504,7 +533,7 @@ describe("monitorTelegramProvider (grammY)", () => {
     const monitor = monitorTelegramProvider({ token: "tok", abortSignal: abort.signal });
     await vi.waitFor(() => expect(runSpy).toHaveBeenCalledTimes(1));
 
-    expect(emitUnhandledRejection(makeTaggedPollingFetchError())).toBe(true);
+    emitUnhandledRejection(makeTaggedPollingFetchError());
     await monitor;
 
     expect(stop.mock.calls.length).toBeGreaterThanOrEqual(1);
@@ -514,6 +543,7 @@ describe("monitorTelegramProvider (grammY)", () => {
   });
 
   it("reuses the resolved transport across polling restarts", async () => {
+    const { monitorTelegramProvider } = await import("./monitor.js");
     vi.useFakeTimers({ shouldAdvanceTime: true });
     try {
       const telegramTransport = {
@@ -542,6 +572,7 @@ describe("monitorTelegramProvider (grammY)", () => {
   });
 
   it("aborts the active Telegram fetch when unhandled network rejection forces restart", async () => {
+    const { monitorTelegramProvider } = await import("./monitor.js");
     const abort = new AbortController();
     const { stop } = mockRunOnceWithStalledPollingRunner();
     mockRunOnceAndAbort(abort);
@@ -552,7 +583,7 @@ describe("monitorTelegramProvider (grammY)", () => {
     expect(firstSignal).toBeInstanceOf(AbortSignal);
     expect((firstSignal as AbortSignal).aborted).toBe(false);
 
-    expect(emitUnhandledRejection(makeTaggedPollingFetchError())).toBe(true);
+    emitUnhandledRejection(makeTaggedPollingFetchError());
     await monitor;
 
     expect((firstSignal as AbortSignal).aborted).toBe(true);
@@ -560,6 +591,7 @@ describe("monitorTelegramProvider (grammY)", () => {
   });
 
   it("ignores unrelated process-level network errors while telegram polling is active", async () => {
+    const { monitorTelegramProvider } = await import("./monitor.js");
     const abort = new AbortController();
     const { stop } = mockRunOnceWithStalledPollingRunner();
 
@@ -585,6 +617,7 @@ describe("monitorTelegramProvider (grammY)", () => {
   });
 
   it("passes configured webhookHost to webhook listener", async () => {
+    const { monitorTelegramProvider } = await import("./monitor.js");
     await monitorTelegramProvider({
       token: "tok",
       useWebhook: true,
@@ -609,6 +642,7 @@ describe("monitorTelegramProvider (grammY)", () => {
   });
 
   it("webhook mode waits for abort signal before returning", async () => {
+    const { monitorTelegramProvider } = await import("./monitor.js");
     const abort = new AbortController();
     const settled = vi.fn();
     const monitor = monitorTelegramProvider({
@@ -628,6 +662,7 @@ describe("monitorTelegramProvider (grammY)", () => {
   });
 
   it("force-restarts polling when getUpdates stalls (watchdog)", async () => {
+    const { monitorTelegramProvider } = await import("./monitor.js");
     vi.useFakeTimers({ shouldAdvanceTime: true });
     const abort = new AbortController();
     const { stop } = mockRunOnceWithStalledPollingRunner();
@@ -668,6 +703,7 @@ describe("monitorTelegramProvider (grammY)", () => {
   });
 
   it("resets webhookCleared latch on 409 conflict so deleteWebhook re-runs", async () => {
+    const { monitorTelegramProvider } = await import("./monitor.js");
     const abort = new AbortController();
     api.deleteWebhook.mockReset();
     api.deleteWebhook.mockResolvedValue(true);
@@ -706,6 +742,7 @@ describe("monitorTelegramProvider (grammY)", () => {
   });
 
   it("falls back to configured webhookSecret when not passed explicitly", async () => {
+    const { monitorTelegramProvider } = await import("./monitor.js");
     await monitorTelegramProvider({
       token: "tok",
       useWebhook: true,
